feat: Add work item collect To evaluate code review effectiveness, we need to check if in deph2 relation tree there are any Bug or Defect.
But is not possible to do this without having this tree stored in a database.
Otherwise, the evaluation will need to work with all of this in memory, making
the process memory expensive and slow.

Then, we are adding a work item crawler service. Given a certain period of time,
we will query for all Bugs and Defects workitems, then, request the related stories.

Also, we change the Pull Request Consumption to also store the work items relation.

When we craw for all work itens, the relation is updated on database automatically.

To support the work item consumption, we add the area path configuration.